pyproject.toml
[project]
name = "a2a-samples-workspace"
version = "0.1.0"
description = "Agent2Agent samples"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.10",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]

[tool.uv.workspace]
members = [
    "samples/python/agents/images",
    "samples/python/agents/medical_Images",
    "demo/ui",
]
